Seperated main.cpp in 12-10-bounce in several files
